+++ src/test/java/transform.java
@@ -1,7 +1,6 @@
 
 
 import java.io.File;
-import java.io.Writer;
 
 import be.baur.sda.DataNode;
 import be.baur.sda.SDA;
@@ -17,14 +16,11 @@ public final class transform {
 		Transform tran = SDT.parse(new File(sdtfile));
 		
 		String sdafile = transform.class.getResource("/example.sda").getFile();
-		TransformContext c = new TransformContext.Builder() //.setWriter(SDT.nullWriter())
-			.setStringParameter("filename", sdafile).build();
-		Writer w = c.getWriter();
+		TransformContext c = new TransformContext.Builder()
+			.setStringParameter("filename", sdafile)
+			.build();
 		
-		w.write("<<\n");
-		DataNode node = tran.execute(c); 
-		SDA.format(w, node);
-		w.write(">>\n");
-		w.flush();
+		DataNode node = tran.execute(c);
+		c.write(SDA.format(node));
 	}
 }