+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 As of this release it is possible to query XML documents using SDT.
 
 - `Removed` SDT namespace prefix and URI declarations from the SDT class.
+- `Removed` TransformContext.getWriter(), replaced with write().
 - `Renamed` sdt.xpath.functions package to sdt.functions.
 - `Renamed` StatementContext.setContextNode() to setXPathContext().
 - Added TransformContext.Builder.setNavigator() and .addNamespace().